Moves notification handler to a better place
