+++ lib/assets/javascripts/cartodb3/editor/layers/analysis-views/analyses-tooltip-error.js
@@ -2,7 +2,7 @@ var $ = require('jquery');
 var cdb = require('cartodb.js');
 var CoreView = require('backbone/core-view');
 var TipsyTooltipView = require('../../../components/tipsy-tooltip-view.js');
-var NotificationErrorMessageHandler = require('./notification-error-message-handler');
+var NotificationErrorMessageHandler = require('../notification-error-message-handler');
 
 module.exports = CoreView.extend({
   initialize: function (opts) {
