+++ lib/assets/javascripts/cartodb3/editor/layers/analysis-views/analysis-notifications.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var Notifier = require('../../../components/notifier/notifier');
-var NotificationErrorMessageHandler = require('./notification-error-message-handler');
+var NotificationErrorMessageHandler = require('../notification-error-message-handler');
 
 var DEFAULT_DELAY = Notifier.DEFAULT_DELAY;
 var STATUS_READY = 'ready';
