+++ lib/assets/javascripts/cartodb3/editor/layers/notification-error-message-handler.js
@@ -1,5 +1,5 @@
 var Messages = require('./notification-error-messages.json');
-var Utils = require('../../../../../javascripts/cartodb3/helpers/utils');
+var Utils = require('../../../../javascripts/cartodb3/helpers/utils');
 
 module.exports = {
   extractErrorFromAnalysisNode: function (analysisNode) {
