+++ lib/assets/test/spec/cartodb3/editor/layers/notification-error-message-handler.spec.js
@@ -1,5 +1,5 @@
 var cdb = require('cartodb.js');
-var NotificationErrorHandler = require('../../../../../javascripts/cartodb3/editor/layers/analysis-views/notification-error-message-handler');
+var NotificationErrorHandler = require('../../../../../javascripts/cartodb3/editor/layers/notification-error-message-handler');
 
 describe('NotificationErrorHandler.extractErrorFromAnalysisNode', function () {
   it('should extract the right error message', function () {
